upgrade cooked keyboard codes to 16 bits
This allows special key codes that don't conflict with latin1 or other
non-ASCII characters.

+++ libraries/TTBAS_LIB/ps22tty.h
@@ -8,8 +8,8 @@
 extern TPS2 pb;
 extern struct ring_buffer kbuf;
 
-uint8_t ICACHE_RAM_ATTR ps2read();
-uint8_t ICACHE_RAM_ATTR ps2peek();
+uint16_t ICACHE_RAM_ATTR ps2read();
+uint16_t ICACHE_RAM_ATTR ps2peek();
 
 static inline bool ICACHE_RAM_ATTR ps2kbhit() {
   if (!pb.available() && rb_is_empty(&kbuf))
